+++ packages/vexide-startup/src/lib.rs
@@ -92,11 +92,15 @@ extern "Rust" {
 /// Sets up the user stack, zeroes the BSS section, and calls the user code.
 /// This function is designed to be used as an entrypoint for programs on the VEX V5 Brain.
 ///
+/// # Const Parameters
+///
+/// - `BANNER`: Enables the vexide startup banner, which prints the vexide logo ASCII art and a startup message.
+///
 /// # Safety
 ///
 /// This function MUST only be called once and should only be called at the very start of program initialization.
 /// Calling this function more than one time will seriously mess up both your stack and your heap.
-pub unsafe fn program_entry() {
+pub unsafe fn program_entry<const BANNER: bool>() {
     #[cfg(target_arch = "arm")]
     unsafe {
         use core::arch::asm;
@@ -127,9 +131,9 @@ pub unsafe fn program_entry() {
         #[cfg(target_arch = "arm")]
         vexide_core::allocator::vexos::init_heap();
         // Print the banner
-        #[cfg(not(feature = "no-banner"))]
-        vexide_core::io::print!(
-            "
+        if BANNER {
+            vexide_core::io::print!(
+                "
 \x1B[1;38;5;196m=%%%%%#-  \x1B[38;5;254m-#%%%%-\x1B[1;38;5;196m  :*%%%%%+.
 \x1B[38;5;208m  -#%%%%#-  \x1B[38;5;254m:%-\x1B[1;38;5;208m  -*%%%%#
 \x1B[38;5;226m    *%%%%#=   -#%%%%%+
@@ -140,8 +144,8 @@ pub unsafe fn program_entry() {
 vexide startup successful!
 Running user code...
 "
-        );
-        vex_sdk::vexTasksRun();
+            );
+        }
         // Run vexos background processing at a regular 2ms interval.
         // This is necessary for serial and device reads to work properly.
         vexide_async::task::spawn(async {